+++ R/estimate_gps.R
@@ -5,11 +5,11 @@
 #' Estimates GPS value for each observation using normal or kernel
 #' approaches.
 #'
-#'
-#' @param w A data frame of observed continuous exposure variable. Including
-#' `id` and `w` columns.
-#' @param c A data frame of observed covariates variable. Also includes `id`
-#' column.
+#' @param data A data frame of observed continuous exposure variable and
+#' observed covariates variable. Also includes `id` column for future
+#' references.
+#' @param formula A formula specifying the relationship between the exposure
+#' variable and the covariates. For example, w ~ I(cf1^2) + cf2.
 #' @param gps_density Model type which is used for estimating GPS value,
 #' including `normal` (default) and `kernel`.
 #' @param params Includes list of parameters that are used internally. Unrelated
@@ -21,13 +21,11 @@
 #'
 #' @return
 #' The function returns a S3 object. Including the following:
-#'   - `dataset `: `id`, `w`, `gps`
-#'   - e_gps_pred
-#'   - e_gps_std_pred
-#'   - w_resid
+#'   - `dataset `: `id`, `w`, `gps`, `e_gps_pred`, `e_gps_std_pred`, `w_resid`
 #'   - gps_mx (min and max of gps)
 #'   - w_mx (min and max of w).
 #'   - used_params
+#'   - formula
 #'
 #' @note
 #' If \code{internal.use} is set to be FALSE, only original data set + GPS will
@@ -42,18 +40,17 @@
 #' @examples
 #' \donttest{
 #' m_d <- generate_syn_data(sample_size = 100)
-#' data_with_gps <- estimate_gps(m_d[, c("id", "w")],
-#'                               m_d[, c("id", "cf1", "cf2", "cf3",
-#'                                       "cf4", "cf5", "cf6")],
+#' data_with_gps <- estimate_gps(data= m_d,
+#'                               formula = w ~ cf1 + cf2 + cf3 + cf4 + cf5 + cf6,
 #'                               gps_density = "normal",
-#'                               params = list(xgb_max_depth = c(3,4,5),
-#'                                        xgb_nrounds=c(10,20,30,40,50,60)),
+#'                               params = list(xgb_max_depth = c(3, 4, 5),
+#'                                        xgb_nrounds=c(10, 20, 30, 40, 50, 60)),
 #'                               nthread = 1,
 #'                               sl_lib = c("m_xgboost")
 #'                              )
 #'}
-estimate_gps <- function(w,
-                         c,
+estimate_gps <- function(data,
+                         formula,
                          gps_density = "normal",
                          params = list(),
                          sl_lib = c("m_xgboost"),
@@ -66,11 +63,8 @@ estimate_gps <- function(w,
   check_args_estimate_gps(gps_density, ...)
 
 
-  id_exist_w <- any(colnames(w) %in% "id")
-  if (!id_exist_w) stop("w should include id column.")
-
-  id_exist_c <- any(colnames(c) %in% "id")
-  if (!id_exist_c) stop("c should include id column.")
+  id_exist <- any(colnames(data) %in% "id")
+  if (!id_exist) stop("data should include id column.")
 
   dot_args <- list(...)
   arg_names <- names(dot_args)
@@ -80,14 +74,19 @@ estimate_gps <- function(w,
   }
 
   # Check if data has missing value(s) -----------------------------------------
-  if (sum(is.na(w)) > 0){
-    logger::log_warn("Vector w has {sum(is.na(w))} missing values.")
+  if (sum(is.na(data)) > 0){
+    logger::log_warn(
+      "data data.frame has {sum(is.na(c))} missing values.")
   }
 
-  if (sum(is.na(c)) > 0){
-    logger::log_warn(
-      "Confounders data.frame (c) has {sum(is.na(c))} missing values.")
-  }
+  # Preprocess the data based on the formula
+  model_data <- model.matrix(object = formula,
+                             data = data)
+
+  response_var = all.vars(formula)[1]
+  response_data = data[[response_var]]
+
+
 
   # Generate SL wrapper library for each type of prediction algorithms ---------
   sl_lib_internal = NULL
@@ -103,44 +102,33 @@ estimate_gps <- function(w,
     }
   }
 
-  merged_data <- merge(w, c, by = "id")
-
-  if (nrow(merged_data) == 0){
-    stop(paste0("Merged data length is 0.",
-                " Make sure that w and c belong to the same observations, ",
-                " or partially include same observations."))
-  }
-
-  exposure_col <- Filter(function(x) !(x %in% c("id")), colnames(w))
-  covariate_cols <- Filter(function(x) !(x %in% c("id")), colnames(c))
-
 
   if (gps_density == "normal"){
-    e_gps <- train_it(target = merged_data[,c(exposure_col)],
-                      input = merged_data[, covariate_cols],
+    e_gps <- train_it(target = response_data,
+                      input = model_data,
                       sl_lib_internal = sl_lib_internal,
                       ...)
 
     e_gps_pred <- e_gps$SL.predict
-    e_gps_std_pred <- stats::sd(merged_data[,c(exposure_col)] - e_gps_pred)
-    w_resid <- compute_resid(merged_data[,c(exposure_col)],
+    e_gps_std_pred <- stats::sd(response_data - e_gps_pred)
+    w_resid <- compute_resid(response_data,
                              e_gps_pred,
                              e_gps_std_pred)
-    gps <- stats::dnorm(merged_data[,c(exposure_col)],
+    gps <- stats::dnorm(response_data,
                         mean = e_gps_pred,
                         sd = e_gps_std_pred)
 
   } else if (gps_density == "kernel"){
 
-    e_gps <- train_it(target = merged_data[,c(exposure_col)],
-                      input = merged_data[, covariate_cols],
+    e_gps <- train_it(target = response_data,
+                      input = model_data,
                       sl_lib_internal = sl_lib_internal, ...)
     e_gps_pred <- e_gps$SL.predict
-    e_gps_std <- train_it(target = abs(merged_data[,c(exposure_col)] - e_gps_pred),
-                          input = merged_data[, covariate_cols],
+    e_gps_std <- train_it(target = abs(response_data - e_gps_pred),
+                          input = model_data,
                           sl_lib_internal = sl_lib_internal, ...)
     e_gps_std_pred <- e_gps_std$SL.predict
-    w_resid <- compute_resid(merged_data[,c(exposure_col)],
+    w_resid <- compute_resid(response_data,
                              e_gps_pred,e_gps_std_pred)
     gps <- compute_density(w_resid, w_resid)
 
@@ -151,13 +139,11 @@ estimate_gps <- function(w,
                ". Use normal or kernel."))
   }
 
-  w_mx <- compute_min_max(merged_data[,c(exposure_col)])
+  w_mx <- compute_min_max(response_data)
   gps_mx <- compute_min_max(gps)
-  merged_data$gps <- gps
 
-  # Drop covariates
-  merged_data[covariate_cols] <- NULL
-  dataset <- merged_data
+  # create new data.frame for output
+  dataset <- data.frame(id = data$id, w = response_data, gps = gps)
   dataset$e_gps_pred <- e_gps_pred
   if (length(e_gps_std_pred) == 1){
     e_gps_std_pred <- rep(e_gps_std_pred, nrow(dataset))
@@ -195,6 +181,7 @@ estimate_gps <- function(w,
   result$used_params <- used_params
   result$gps_mx <- gps_mx
   result$w_mx <- w_mx
+  result$formula <- formula
 
   invisible(result)
 }